+++ src/player/main/sp/it/pl/plugin/impl/WallpaperChanger.kt
@@ -1,23 +1,13 @@
 package sp.it.pl.plugin.impl
 
 import java.io.File
-import javafx.scene.Scene
 import javafx.scene.image.Image
-import javafx.scene.layout.Pane
 import javafx.stage.Screen
-import javafx.stage.Stage
-import javafx.stage.StageStyle.UNDECORATED
-import javafx.stage.WindowEvent.WINDOW_HIDDEN
 import sp.it.pl.image.ImageStandardLoader
 import sp.it.pl.main.APP
-import sp.it.pl.main.Events.AppEvent.SystemSleepEvent
-import sp.it.pl.main.Events.AppEvent.UserSessionEvent
 import sp.it.pl.main.isImage
 import sp.it.pl.plugin.PluginBase
 import sp.it.pl.plugin.PluginInfo
-import sp.it.pl.ui.objects.image.Thumbnail
-import sp.it.pl.ui.objects.window.ShowArea
-import sp.it.pl.ui.objects.window.stage.setNonInteractingProgmanOnBottom
 import sp.it.util.access.readOnly
 import sp.it.util.access.vn
 import sp.it.util.async.runIO
@@ -25,29 +15,19 @@ import sp.it.util.conf.cvn
 import sp.it.util.conf.def
 import sp.it.util.conf.only
 import sp.it.util.file.FileType.FILE
-import sp.it.util.functional.asIf
-import sp.it.util.reactive.Disposer
 import sp.it.util.reactive.Subscribed
-import sp.it.util.reactive.Subscription
-import sp.it.util.reactive.on
-import sp.it.util.reactive.onChangeAndNow
-import sp.it.util.reactive.onEventDown
-import sp.it.util.reactive.sync
-import sp.it.util.reactive.sync1IfNonNull
 import sp.it.util.system.Os
-import sp.it.util.ui.anchorPane
-import sp.it.util.ui.image.FitFrom
+import sp.it.util.system.Windows
+import sp.it.util.ui.image.FitFrom.OUTSIDE
 import sp.it.util.ui.image.ImageSize
-import sp.it.util.ui.lay
-import sp.it.util.ui.size
-import sp.it.util.ui.x
-import sp.it.util.ui.xy
+import sp.it.util.units.uuid
 
 class WallpaperChanger: PluginBase() {
 
    private val wallpaperImageW = vn<Image>(null)
    val wallpaperImage = wallpaperImageW.readOnly()
    val wallpaperFile by cvn<File>(null).only(FILE).def(name = "Wallpaper file") sync ::load
+   val cacheId = uuid("e37ac005-4be3-4241-b81d-ba19dc376857")
    val menuItemInjector = Subscribed {
       APP.contextMenus.menuItemBuilders.add<File> {
          if (value.isImage()) {
@@ -57,83 +37,34 @@ class WallpaperChanger: PluginBase() {
          }
       }
    }
-   private val wallpaperApplier = Subscribed {
-      val disposer = Disposer()
-
-      // An all-screen window displaying the wallpaper image in a positioned thumbnail per each screen
-      Stage().run {
-         val root = anchorPane()
-         val b = ShowArea.ALL_SCREENS.bounds().second
-
-         initStyle(UNDECORATED)
-         title = "${APP.name}-Wallpaper"
-         scene = Scene(root)
-         xy = 0 x 0
-         size = b.size
-         setNonInteractingProgmanOnBottom()
-
-         Screen.getScreens().forEach { screen ->
-            root.lay(screen.bounds.minY - b.minY, null, null, screen.bounds.minX - b.minX) += Thumbnail(screen.bounds.size).run {
-               fitFrom.value = FitFrom.OUTSIDE
-               wallpaperImageW sync { if (it!=null) loadImage(it) } on disposer
-               onEventDown(WINDOW_HIDDEN) { loadImage(null) }
-
-               pane
-            }
-         }
-
-         disposer += {
-            close()
-            scene.root.asIf<Pane>()?.children?.clear()
-            scene = null
-         }
-
-         wallpaperImageW.sync1IfNonNull { show() } on disposer
-      }
-
-      Subscription { disposer() }
-   }
-   private val overlaySleepHandler = Subscribed {
-      APP.actionStream.onEventObject(SystemSleepEvent.Stop) { wallpaperApplier.resubscribe() }
-   }
-   private val overlayUserHandler = Subscribed {
-      APP.actionStream.onEventObject(UserSessionEvent.Start) { wallpaperApplier.resubscribe() }
-   }
-   private val wallpaperIsShowing = Subscribed {
-      APP.mouse.screens.onChangeAndNow { wallpaperApplier.resubscribe() }
-   }
 
    private fun load(f: File?) {
+      val size = largestScreenSize()
       runIO {
-         ImageStandardLoader(f, largestScreenSize(), true)
-      } ui { image ->
-         wallpaperImageW.value = image
+         if (f!=null) Windows.changeWallpaper(f)
+         ImageStandardLoader.memoized(cacheId)(f, size, OUTSIDE, true)
+      } ui {
+         wallpaperImageW.value = it
       }
    }
 
    override fun start() {
-      wallpaperIsShowing.subscribe()
       menuItemInjector.subscribe()
-      overlaySleepHandler.subscribe()
-      overlayUserHandler.subscribe()
    }
 
    override fun stop() {
-      overlaySleepHandler.unsubscribe()
-      overlayUserHandler.unsubscribe()
       menuItemInjector.unsubscribe()
-      wallpaperIsShowing.unsubscribe()
       wallpaperImageW.value = null
    }
 
    companion object: PluginInfo {
       override val name = "Wallpaper"
-      override val description = "Provides the ability to change wallpaper until OS shutdown. Also improves screen overlay effect performance."
+      override val description = "Provides the ability to change wallpaper. Also improves screen overlay loading speed."
       override val isSupported = Os.WINDOWS.isCurrent
       override val isSingleton = true
       override val isEnabledByDefault = false
 
-      fun largestScreenSize() = Screen.getScreens().map { ImageSize(it.bounds.width, it.bounds.height) }.maxByOrNull { it.width*it.height }
+      fun largestScreenSize() = Screen.getScreens().map { ImageSize(it.bounds.width, it.bounds.height) }.maxByOrNull { it.width*it.height }?.div(2.0)
          ?: ImageSize(0.0, 0.0)
    }
 }